+++ wxcadm/person.py
@@ -11,7 +11,7 @@ from collections import UserList
 import wxcadm.exceptions
 from .common import *
 from .xsi import XSI
-from .device import DeviceList, Device
+from .device import DeviceList, Device, SupportedDevice, SupportedDeviceList
 from .location import Location
 
 from wxcadm import log
@@ -588,94 +588,6 @@ class Person:
         if self._devices is None:
             self._devices = DeviceList(self)
         return self._devices
-
-    def add_device(self, model: str, mac: Optional[str] = None, password: Optional[str] = None):
-        """ Add a new device to the Person
-
-        In order to use this method, you must know the model of the device that you are adding, as expected by the
-        Webex API. If you are adding a "Generic IPPhone Customer Managed" device, you can use that value or simply
-        send ``model='GENERIC'`` as an alias. You can find the full list of models with the
-        :py:meth:`Org.get_supported_devices()` method.
-
-        If the MAC address is passed, a device will be created with the provided MAC address. If no MAC address is
-        passed, an Activation Code will be generated and returned as part of the response. Your integration/token must
-        have the ``identity:placeonetimepassword_create`` scope to create Activation Codes for devices.
-
-        .. deprecated:: 4.3.0
-            Use :meth:`Person.devices.create()` instead
-
-        Args:
-            model (str): The model name of the device being added
-            mac (str, optional): The MAC address of the device being added
-            password (str, optional): Only valid when creating a Generic IPPhone Customer Managed device. If a
-                password is not provided, the Webex API will generate a unique, compliant SIP password and return it
-                in the response.
-
-        Returns:
-            dict: The dict values vary based on the type of device being activated. If the device fails for any reason,
-                False will be returned. At the moment, Webex doesn't provide very useful failure reasons, but those may
-                be added to the return value when they are available.
-
-        """
-        log.info(f"Adding a device to {self.display_name}")
-        payload = {
-            "personId": self.id,
-            "model": model
-        }
-        data_needed = False  # Flag that we need to get platform data once we have a Device ID
-        if mac is None:
-            # If no MAC address is provided, just generate an activation code for the device
-            try:
-                response = webex_api_call('post',
-                                          '/v1/devices/activationCode',
-                                          payload=payload,
-                                          params={'orgId': self._parent.id})
-                log.debug(f"\t{response}")
-            except wxcadm.exceptions.APIError:
-                return False
-
-            # Get the ID of the device we just inserted
-            device_id = response.get('id', None).replace('=', '')
-
-            results = {
-                'device_id': device_id,
-                'activation_code': response['code']
-            }
-        else:
-            payload['mac'] = mac
-            if model.upper() == "GENERIC" or model.upper() == "Generic IPPhone Customer Managed":
-                payload['model'] = "Generic IPPhone Customer Managed"  # Hard-code what the API expects (for now)
-                data_needed = True
-                if password is None:  # Generate a unique password
-                    response = webex_api_call('POST',
-                                              f'v1/telephony/config/locations/{self.location}/actions/'
-                                              f'generatePassword/invoke')
-                    password = response['exampleSipPassword']
-                payload['password'] = password
-            response = webex_api_call('post', 'v1/devices', payload=payload, params={'orgId': self._parent.id})
-            log.debug(f"\t{response}")
-
-            # Get the ID of the device we just inserted
-            device_id = response.get('id', None).replace('=', '')
-
-            results = {
-                'device_id': device_id,
-                'mac': response['mac']
-            }
-
-            if data_needed is True:
-                response = webex_api_call('get', f'/v1/telephony/config/devices/{device_id}')
-                results['sip_auth_user'] = response['owner']['sipUserName']
-                results['line_port'] = response['owner']['linePort']
-                results['password'] = password
-                results['sip_userpart'] = response['owner']['linePort'].split('@')[0]
-                results['sip_hostpart'] = response['owner']['linePort'].split('@')[1]
-                results['sip_outbound_proxy'] = response['proxy']['outboundProxy']
-                results['sip_outbound_proxy_srv'] = f"_sips._tcp.{response['proxy']['outboundProxy']}"
-
-        # Provide the Device instance in the response as well
-        results['device_object'] = Device(parent=self, id=device_id)
-        return results
 
     @property
     def hunt_groups(self):